works host FreeBSD 13 php74

+++ bitrix/modules/main/classes/mysql/user_counter.php
@@ -1,4 +1,4 @@
-<?
+<?php
 require_once($_SERVER["DOCUMENT_ROOT"]."/bitrix/modules/main/classes/general/user_counter.php");
 
 class CUserCounter extends CAllUserCounter
@@ -9,7 +9,7 @@ class CUserCounter extends CAllUserCounter
 
 		$value = intval($value);
 		$user_id = intval($user_id);
-		if ($user_id < 0 || strlen($code) <= 0)
+		if ($user_id < 0 || $code == '')
 			return false;
 
 		$rs = $DB->Query("
@@ -19,19 +19,27 @@ class CUserCounter extends CAllUserCounter
 			AND CODE = '".$DB->ForSQL($code)."'
 		");
 
-		if ($rs->Fetch())
+		if ($cntVal = $rs->Fetch())
 		{
 			$ssql = "";
 			if ($tag != "")
 				$ssql = ", TAG = '".$DB->ForSQL($tag)."'";
 
-			$DB->Query("
-				UPDATE b_user_counter SET
-				CNT = ".$value." ".$ssql."
-				WHERE USER_ID = ".$user_id."
-				AND SITE_ID = '".$DB->ForSQL($site_id)."'
-				AND CODE = '".$DB->ForSQL($code)."'
-			");
+			if($cntVal['CNT'] != $value)
+			{
+				$DB->Query("
+					UPDATE b_user_counter SET
+					CNT = " . $value . " " . $ssql . ",
+					SENT = 0
+					WHERE USER_ID = " . $user_id . "
+					AND SITE_ID = '" . $DB->ForSQL($site_id) . "'
+					AND CODE = '" . $DB->ForSQL($code) . "'
+				");
+			}
+			else
+			{
+				$sendPull = false;
+			}
 		}
 		else
 		{
@@ -74,7 +82,7 @@ class CUserCounter extends CAllUserCounter
 		global $DB, $CACHE_MANAGER;
 
 		$user_id = intval($user_id);
-		if ($user_id < 0 || strlen($code) <= 0)
+		if ($user_id < 0 || $code == '')
 			return false;
 
 		$increment = intval($increment);
@@ -117,14 +125,20 @@ class CUserCounter extends CAllUserCounter
 	}
 
 	/**
-	* @deprecated
-	*/
+	 * @deprecated
+	 * @param $user_id
+	 * @param $code
+	 * @param string $site_id
+	 * @param bool $sendPull
+	 * @param int $decrement
+	 * @return bool
+	 */
 	public static function Decrement($user_id, $code, $site_id = SITE_ID, $sendPull = true, $decrement = 1)
 	{
 		global $DB, $CACHE_MANAGER;
 
 		$user_id = intval($user_id);
-		if ($user_id < 0 || strlen($code) <= 0)
+		if ($user_id < 0 || $code == '')
 			return false;
 
 		$decrement = intval($decrement);
@@ -171,7 +185,7 @@ class CUserCounter extends CAllUserCounter
 	{
 		global $DB, $CACHE_MANAGER, $APPLICATION;
 
-		if (strlen($sub_select) > 0)
+		if ($sub_select <> '')
 		{
 			$pullInclude = (
 				$sendPull
@@ -184,7 +198,7 @@ class CUserCounter extends CAllUserCounter
 			)
 			{
 				$strSQL = "
-					INSERT INTO b_user_counter (USER_ID, CNT, SITE_ID, CODE, SENT, TAG) (".$sub_select.")
+					INSERT INTO b_user_counter (USER_ID, CNT, SITE_ID, CODE, SENT, TAG".(is_array($arParams) && isset($arParams["SET_TIMESTAMP"]) ? ", TIMESTAMP_X" : "").") (".$sub_select.")
 					ON DUPLICATE KEY UPDATE CNT = CNT + VALUES(CNT), SENT = VALUES(SENT), TAG = '".$DB->ForSQL($arParams["TAG_SET"])."'
 				";
 			}
@@ -194,7 +208,7 @@ class CUserCounter extends CAllUserCounter
 			)
 			{
 				$strSQL = "
-					INSERT INTO b_user_counter (USER_ID, CNT, SITE_ID, CODE, SENT) (".$sub_select.")
+					INSERT INTO b_user_counter (USER_ID, CNT, SITE_ID, CODE, SENT".(is_array($arParams) && isset($arParams["SET_TIMESTAMP"]) ? ", TIMESTAMP_X" : "").") (".$sub_select.")
 					ON DUPLICATE KEY UPDATE CNT = CASE
 						WHEN
 							TAG = '".$DB->ForSQL($arParams["TAG_CHECK"])."'
@@ -217,7 +231,7 @@ class CUserCounter extends CAllUserCounter
 			else
 			{
 				$strSQL = "
-					INSERT INTO b_user_counter (USER_ID, CNT, SITE_ID, CODE, SENT) (".$sub_select.")
+					INSERT INTO b_user_counter (USER_ID, CNT, SITE_ID, CODE, SENT".(is_array($arParams) && isset($arParams["SET_TIMESTAMP"]) ? ", TIMESTAMP_X" : "").") (".$sub_select.")
 					ON DUPLICATE KEY UPDATE CNT = CNT + VALUES(CNT), SENT = VALUES(SENT)
 				";
 			}
@@ -241,51 +255,62 @@ class CUserCounter extends CAllUserCounter
 
 			if ($pullInclude)
 			{
-				$db_lock = $DB->Query("SELECT GET_LOCK('".$APPLICATION->GetServerUniqID()."_pull', 0) as L");
-				$ar_lock = $db_lock->Fetch();
-				if($ar_lock["L"] > 0)
+				$arSites = Array();
+				$by = '';
+				$order = '';
+				$res = CSite::GetList($by, $order, Array("ACTIVE" => "Y"));
+				while($row = $res->Fetch())
 				{
-					$arSites = Array();
-					$res = CSite::GetList(($b = ""), ($o = ""), Array("ACTIVE" => "Y"));
-					while($row = $res->Fetch())
+					$arSites[] = $row['ID'];
+				}
+
+				if (
+					!empty($arParams["USERS_TO_PUSH"])
+					&& is_array($arParams["USERS_TO_PUSH"])
+				)
+				{
+					$connection = \Bitrix\Main\Application::getConnection();
+					if($connection->lock('pull'))
 					{
-						$arSites[] = $row['ID'];
+						$helper = $connection->getSqlHelper();
+
+						$strSQL = "
+							SELECT pc.CHANNEL_ID, uc.USER_ID, uc.SITE_ID, uc.CODE, uc.CNT
+							FROM b_user_counter uc
+							INNER JOIN b_pull_channel pc ON pc.USER_ID = uc.USER_ID
+							INNER JOIN b_user u ON u.ID = uc.USER_ID AND (CASE WHEN u.EXTERNAL_AUTH_ID IN ('".join("', '", \Bitrix\Main\UserTable::getExternalUserTypes())."') THEN 'Y' ELSE 'N' END) = 'N' AND u.LAST_ACTIVITY_DATE > ".$helper->addSecondsToDateTime('(-3600)')."
+							WHERE uc.SENT = '0' AND uc.USER_ID IN (".implode(", ", $arParams["USERS_TO_PUSH"]).")
+						";
+
+						$res = $DB->Query($strSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
+
+						$pullMessage = Array();
+						while($row = $res->Fetch())
+						{
+							CUserCounter::addValueToPullMessage($row, $arSites, $pullMessage);
+						}
+
+						$DB->Query("UPDATE b_user_counter SET SENT = '1' WHERE SENT = '0' AND CODE NOT LIKE '".CUserCounter::LIVEFEED_CODE."L%'");
+
+						$connection->unlock('pull');
+
+						if (\CUserCounter::CheckLiveMode())
+						{
+							foreach ($pullMessage as $channelId => $arMessage)
+							{
+								\Bitrix\Pull\Event::add($channelId, Array(
+									'module_id' => 'main',
+									'command' => 'user_counter',
+									'expiry' => 3600,
+									'params' => $arMessage,
+								));
+							}
+						}
 					}
-
-					$strUserFilter = (
-						!empty($arParams["USERS_TO_PUSH"])
-						&& is_array($arParams["USERS_TO_PUSH"])
-							? " AND uc.USER_ID IN (".implode(", ", $arParams["USERS_TO_PUSH"]).")"
-							: ""
-					);
-
-					$strSQL = "
-						SELECT pc.CHANNEL_ID, uc.USER_ID, uc.SITE_ID, uc.CODE, uc.CNT
-						FROM b_user_counter uc
-						INNER JOIN b_pull_channel pc ON pc.USER_ID = uc.USER_ID
-						WHERE uc.SENT = '0'".$strUserFilter."
-					";
-
-					$res = $DB->Query($strSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
-
-					$pullMessage = Array();
-					while($row = $res->Fetch())
-					{
-						CUserCounter::addValueToPullMessage($row, $arSites, $pullMessage);
-					}
-
-					$DB->Query("UPDATE b_user_counter SET SENT = '1' WHERE SENT = '0' AND CODE NOT LIKE '**L%'");
-					$DB->Query("SELECT RELEASE_LOCK('".$APPLICATION->GetServerUniqID()."_pull')");
-
-					foreach ($pullMessage as $channelId => $arMessage)
-					{
-						CPullStack::AddByChannel($channelId, Array(
-							'module_id' => 'main',
-							'command' => 'user_counter',
-							'expiry' 	=> 3600,
-							'params' => $arMessage,
-						));
-					}
+				}
+				else
+				{
+					CUserCounterPage::setNewEvent();
 				}
 			}
 		}
@@ -293,12 +318,12 @@ class CUserCounter extends CAllUserCounter
 
 	public static function Clear($user_id, $code, $site_id = SITE_ID, $sendPull = true, $bMultiple = false)
 	{
-		global $DB, $CACHE_MANAGER;
+		global $DB, $CACHE_MANAGER, $APPLICATION;
 
 		$user_id = intval($user_id);
 		if (
 			$user_id < 0
-			|| strlen($code) <= 0
+			|| $code == ''
 		)
 		{
 			return false;
@@ -311,37 +336,43 @@ class CUserCounter extends CAllUserCounter
 
 		if ($bMultiple)
 		{
-			$siteToDelete = "";
-			$strUpsertSQL = "
-				INSERT INTO b_user_counter (USER_ID, SITE_ID, CODE, CNT, LAST_DATE) VALUES ";
-
-			foreach ($site_id as $i => $site_id_tmp)
+			$connection = \Bitrix\Main\Application::getConnection();
+			if($connection->lock('counter_delete'))
 			{
-				if ($i > 0)
+				$siteToDelete = "";
+				$strUpsertSQL = "
+					INSERT INTO b_user_counter (USER_ID, SITE_ID, CODE, CNT, LAST_DATE) VALUES ";
+
+				foreach ($site_id as $i => $site_id_tmp)
 				{
-					$strUpsertSQL .= ",";
-					$siteToDelete .= ",";
+					if ($i > 0)
+					{
+						$strUpsertSQL .= ",";
+						$siteToDelete .= ",";
+					}
+
+					$siteToDelete .= "'".$DB->ForSQL($site_id_tmp)."'";
+					$strUpsertSQL .= " (".$user_id.", '".$DB->ForSQL($site_id_tmp)."', '".$DB->ForSQL($code)."', 0, ".$DB->CurrentTimeFunction().") ";
 				}
+				$strUpsertSQL .= " ON DUPLICATE KEY UPDATE CNT = 0, LAST_DATE = ".$DB->CurrentTimeFunction();
 
-				$siteToDelete .= "'".$DB->ForSQL($site_id_tmp)."'";
-				$strUpsertSQL .= " (".$user_id.", '".$DB->ForSQL($site_id_tmp)."', '".$DB->ForSQL($code)."', 0, ".$DB->CurrentTimeFunction().") ";
+				$strDeleteSQL = "
+					DELETE FROM b_user_counter
+					WHERE
+						USER_ID = ".$user_id."
+						".(
+							count($site_id) == 1
+								? " AND SITE_ID = '".$site_id[0]."' "
+								: " AND SITE_ID IN (".$siteToDelete.") "
+						)."
+						AND CODE LIKE '".$DB->ForSQL($code)."L%'
+					";
+
+				$DB->Query($strDeleteSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
+				$DB->Query($strUpsertSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
+
+				$connection->unlock('counter_delete');
 			}
-			$strUpsertSQL .= " ON DUPLICATE KEY UPDATE CNT = 0, LAST_DATE = ".$DB->CurrentTimeFunction();
-
-			$strDeleteSQL = "
-				DELETE FROM b_user_counter
-				WHERE
-					USER_ID = ".$user_id."
-					".(
-						count($site_id) == 1
-							? " AND SITE_ID = '".$site_id[0]."' "
-							: " AND SITE_ID IN (".$siteToDelete.") "
-					)."
-					AND CODE LIKE '".$DB->ForSQL($code)."L%'
-				";
-
-			$DB->Query($strDeleteSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
-			$DB->Query($strUpsertSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
 		}
 		else
 		{
@@ -357,7 +388,7 @@ class CUserCounter extends CAllUserCounter
 
 			$strSQL .= " ON DUPLICATE KEY UPDATE CNT = 0, LAST_DATE = ".$DB->CurrentTimeFunction();
 
-			$res = $DB->Query($strSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
+			$DB->Query($strSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
 		}
 
 		if (self::$counters && self::$counters[$user_id])
@@ -391,7 +422,7 @@ class CUserCounter extends CAllUserCounter
 	{
 		global $DB, $APPLICATION, $CACHE_MANAGER;
 
-		if (strlen($code) <= 0)
+		if ($code == '')
 		{
 			return false;
 		}
@@ -399,33 +430,41 @@ class CUserCounter extends CAllUserCounter
 		$pullMessage = Array();
 		$bPullEnabled = false;
 
+		$connection = \Bitrix\Main\Application::getConnection();
+
 		if (self::CheckLiveMode())
 		{
-			$db_lock = $DB->Query("SELECT GET_LOCK('".$APPLICATION->GetServerUniqID()."_pull', 0) as L");
-			$ar_lock = $db_lock->Fetch();
-			if ($ar_lock["L"] > 0)
+			if ($connection->lock('pull'))
 			{
 				$bPullEnabled = true;
 
-				$arSites = Array();
-				$res = CSite::GetList(($b = ""), ($o = ""), Array("ACTIVE" => "Y"));
+				$arSites = array();
+				$by = '';
+				$order = '';
+				$res = CSite::GetList($by, $order, array("ACTIVE" => "Y"));
 				while($row = $res->Fetch())
 				{
 					$arSites[] = $row['ID'];
 				}
 
+				$isLF = (
+					mb_substr($code, 0, 2) == CUserCounter::LIVEFEED_CODE
+					&& $code != CUserCounter::LIVEFEED_CODE
+				);
+
+				$helper = $connection->getSqlHelper();
 				$strSQL = "
 					SELECT pc.CHANNEL_ID, uc.USER_ID, uc.SITE_ID, uc.CODE, uc.CNT
 					FROM b_user_counter uc
 					INNER JOIN b_pull_channel pc ON pc.USER_ID = uc.USER_ID
-					WHERE uc.CODE LIKE '**%'
-				";
+					INNER JOIN b_user u ON u.ID = uc.USER_ID AND (CASE WHEN u.EXTERNAL_AUTH_ID IN ('".join("', '", \Bitrix\Main\UserTable::getExternalUserTypes())."') THEN 'Y' ELSE 'N' END) = 'N' AND u.LAST_ACTIVITY_DATE > ".$helper->addSecondsToDateTime('(-3600)')."
+					WHERE uc.CODE ".($isLF ? " LIKE '".CUserCounter::LIVEFEED_CODE."%'" : " = '".$code."'");
 
 				$res = $DB->Query($strSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
 
 				while($row = $res->Fetch())
 				{
-					if ($row["CODE"] == $code)
+					if ($isLF && $row["CODE"] == $code)
 					{
 						continue;
 					}
@@ -442,18 +481,23 @@ class CUserCounter extends CAllUserCounter
 
 		if ($bPullEnabled)
 		{
-			$DB->Query("SELECT RELEASE_LOCK('".$APPLICATION->GetServerUniqID()."_pull')");
+			$connection->unlock('pull');
 		}
 
-		foreach ($pullMessage as $channelId => $arMessage)
+		if (\CUserCounter::CheckLiveMode())
 		{
-			CPullStack::AddByChannel($channelId, Array(
-				'module_id' => 'main',
-				'command' => 'user_counter',
-				'expiry' 	=> 3600,
-				'params' => $arMessage,
-			));
+			foreach ($pullMessage as $channelId => $arMessage)
+			{
+				\Bitrix\Pull\Event::add($channelId, Array(
+					'module_id' => 'main',
+					'command' => 'user_counter',
+					'expiry' 	=> 3600,
+					'params' => $arMessage,
+				));
+			}
 		}
+
+		return null;
 	}
 
 	protected static function dbIF($condition, $yes, $no)
@@ -462,9 +506,111 @@ class CUserCounter extends CAllUserCounter
 	}
 
 	// legacy function
-	public static function ClearByUser($user_id, $site_id = SITE_ID, $code = self::ALL_SITES, $bMultiple = false)
+	public static function ClearByUser($user_id, $site_id = SITE_ID, $code = self::ALL_SITES, $bMultiple = false, $sendPull = true)
 	{
-		return self::Clear($user_id, $code, $site_id, true, $bMultiple);
+		return self::Clear($user_id, $code, $site_id, $sendPull, $bMultiple);
 	}
 }
-?>
+
+class CUserCounterPage extends CAllUserCounterPage
+{
+	public static function checkSendCounter()
+	{
+		global $DB, $USER;
+
+		$connection = \Bitrix\Main\Application::getConnection();
+
+		if(!$connection->lock('counterpull'))
+		{
+			return;
+		}
+
+		$counterPageSize = intval(CAllUserCounterPage::getPageSizeOption(100));
+
+		$userSQL = "SELECT USER_ID FROM b_user_counter WHERE SENT='0' GROUP BY USER_ID LIMIT ".$counterPageSize;
+		$res = $DB->Query($userSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
+
+		$pullMessage = array();
+		$userIdList = array();
+
+		while($row = $res->Fetch())
+		{
+			$userIdList[] = intval($row["USER_ID"]);
+		}
+
+		if (
+			count($userIdList) >= $counterPageSize
+			&& is_object($USER)
+			&& $USER->isAuthorized()
+			&& !in_array($USER->getId(), $userIdList)
+		)
+		{
+			$userIdList[] = $USER->getId();
+		}
+
+		$userString = '';
+		foreach($userIdList as $userId)
+		{
+			$userString .= ($userString <> ''? ', ' : '').$userId;
+		}
+
+		if ($userString <> '')
+		{
+			$arSites = array();
+			$by = '';
+			$order = '';
+			$res = CSite::GetList($by, $order, array("ACTIVE" => "Y"));
+			while($row = $res->Fetch())
+			{
+				$arSites[] = $row['ID'];
+			}
+
+			$helper = $connection->getSqlHelper();
+
+			$strSQL = "
+				SELECT pc.CHANNEL_ID, uc.USER_ID, uc.SITE_ID, uc.CODE, uc.CNT
+				FROM b_user_counter uc
+				INNER JOIN b_pull_channel pc ON pc.USER_ID = uc.USER_ID
+				INNER JOIN b_user u ON u.ID = uc.USER_ID AND (CASE WHEN u.EXTERNAL_AUTH_ID IN ('".join("', '", \Bitrix\Main\UserTable::getExternalUserTypes())."') THEN 'Y' ELSE 'N' END) = 'N' AND u.LAST_ACTIVITY_DATE > ".$helper->addSecondsToDateTime('(-3600)')."
+				WHERE uc.USER_ID IN (".$userString.") AND uc.CODE NOT LIKE '".CUserCounter::LIVEFEED_CODE."L%' AND uc.SENT = '0'
+			";
+
+			$res = $DB->Query($strSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
+			while($row = $res->Fetch())
+			{
+				CUserCounter::addValueToPullMessage($row, $arSites, $pullMessage);
+			}
+
+			$strSQL = "
+				SELECT pc.CHANNEL_ID, uc.USER_ID, uc.SITE_ID, uc.CODE, uc.CNT
+				FROM b_user_counter uc
+				INNER JOIN b_pull_channel pc ON pc.USER_ID = uc.USER_ID
+				INNER JOIN b_user u ON u.ID = uc.USER_ID AND (CASE WHEN u.EXTERNAL_AUTH_ID IN ('".join("', '", \Bitrix\Main\UserTable::getExternalUserTypes())."') THEN 'Y' ELSE 'N' END) = 'N' AND u.LAST_ACTIVITY_DATE > ".$helper->addSecondsToDateTime('(-3600)')."
+				WHERE uc.USER_ID IN (".$userString.") AND uc.CODE LIKE '".CUserCounter::LIVEFEED_CODE."L%'
+			";
+
+			$res = $DB->Query($strSQL, false, "FILE: ".__FILE__."<br> LINE: ".__LINE__);
+			while($row = $res->Fetch())
+			{
+				CUserCounter::addValueToPullMessage($row, $arSites, $pullMessage);
+			}
+
+			$DB->Query("UPDATE b_user_counter SET SENT = '1' WHERE SENT = '0' AND USER_ID IN (".$userString.")");
+		}
+
+		$connection->unlock('counterpull');
+
+		if (\CUserCounter::CheckLiveMode())
+		{
+			foreach ($pullMessage as $channelId => $arMessage)
+			{
+				\Bitrix\Pull\Event::add($channelId, Array(
+					'module_id' => 'main',
+					'command' => 'user_counter',
+					'expiry' => 3600,
+					'params' => $arMessage,
+				));
+			}
+		}
+	}
+}
