+++ bitrix/modules/sale/admin/company.php
@@ -4,7 +4,8 @@ use \Bitrix\Main\Application;
 use Bitrix\Main\Localization\Loc;
 
 require_once($_SERVER["DOCUMENT_ROOT"]."/bitrix/modules/main/include/prolog_admin_before.php");
-require_once($_SERVER["DOCUMENT_ROOT"]."/bitrix/modules/sale/include.php");
+
+\Bitrix\Main\Loader::includeModule('sale');
 
 $saleModulePermissions = $APPLICATION->GetGroupRight("sale");
 if ($saleModulePermissions < "U")
@@ -33,13 +34,13 @@ $USER_FIELD_MANAGER->AdminListAddFilterFields($entity_id, $filterFields);
 $lAdmin->InitFilter($filterFields);
 $filter = array();
 
-if (strlen($filter_name) > 0)
+if ($filter_name <> '')
 	$filter["?NAME"] = $filter_name;
 
-if (strlen($filter_location_id) > 0)
+if ($filter_location_id <> '')
 	$filter["LOCATION_ID"] = $filter_location_id;
 
-if (strlen($filter_active) > 0 && $filter_active != 'NOT_REF')
+if ($filter_active <> '' && $filter_active != 'NOT_REF')
 	$filter["ACTIVE"] = $filter_active;
 
 $USER_FIELD_MANAGER->AdminListAddFilter(CompanyTable::getUfId(), $filter);
@@ -107,17 +108,29 @@ if (($ids = $lAdmin->GroupAction()) && $saleModulePermissions >= "W")
 			case "delete":
 				@set_time_limit(0);
 
-				$conn->startTransaction();
+				$dbRes = \Bitrix\Sale\Internals\OrderTable::getList(array(
+					'select' => array('ID'),
+					'filter' => array(
+						'LOGIC' => 'OR',
+						'SHIPMENT.COMPANY_ID' => $id,
+						'PAYMENT.COMPANY_ID' => $id
+					)
+				));
+
+				if ($dbRes->fetch())
+				{
+					$lAdmin->AddGroupError(Loc::getMessage("SALE_COMPANY_ERROR_DELETE_LINK"), $id);
+					continue 2;
+				}
+
 				$result = CompanyTable::delete($id);
 				if (!$result->isSuccess())
 				{
-					$conn->rollbackTransaction();
 					if ($error = $result->getErrorMessages())
 						$lAdmin->AddGroupError(join("\n", $error), $id);
 					else
 						$lAdmin->AddGroupError(Loc::getMessage("SALE_COMPANY_ERROR_DELETE"), $id);
 				}
-				$conn->commitTransaction();
 				break;
 		}
 	}
@@ -170,6 +183,12 @@ $headers = array(
 		"content"=>Loc::getMessage("SALE_COMPANY_CODE"),
 		"sort"=>"CODE",
 		"default"=>true
+	),
+	array(
+		"id"=>"SORT",
+		"content"=>Loc::getMessage("SALE_COMPANY_SORT"),
+		"sort"=>"SORT",
+		"default"=>true
 	)
 );
 $USER_FIELD_MANAGER->AdminListAddHeaders(CompanyTable::getUfId(), $headers);
@@ -188,19 +207,26 @@ $allSelectedFields = array_merge($allSelectedFields, array_fill_keys($selectedFi
 
 while ($company = $dbResultList->NavNext(true, "f_"))
 {
-	$res = \Bitrix\Sale\Location\LocationTable::getPathToNodeByCode(
-		$company['LOCATION_ID'],
-		array(
-			'select' => array('CHAIN' => 'NAME.NAME'),
-			'filter' => array('NAME.LANGUAGE_ID' => Application::getInstance()->getContext()->getLanguage())
-		)
-	);
-	$path = array();
-	while($item = $res->fetch())
-	    $path[] = $item['CHAIN'];
+	try
+	{
+		$res = \Bitrix\Sale\Location\LocationTable::getPathToNodeByCode(
+				$company['LOCATION_ID'],
+				array(
+						'select' => array('CHAIN' => 'NAME.NAME'),
+						'filter' => array('NAME.LANGUAGE_ID' => Application::getInstance()->getContext()->getLanguage())
+				)
+		);
 
-	$company['LOCATION_ID'] = implode(', ', array_reverse($path));
+		$path = array();
+		while ($item = $res->fetch())
+			$path[] = $item['CHAIN'];
 
+		$company['LOCATION_ID'] = implode(', ', array_reverse($path));
+	}
+	catch (\Bitrix\Main\SystemException $e)
+	{
+		$company['LOCATION_ID'] = '';
+	}
 
 	$row = &$lAdmin->AddRow($f_ID, $company, "sale_company_edit.php?ID=".$f_ID."&lang=".$lang, Loc::getMessage("SALE_COMPANY_EDIT_DESCR"));
